password/requirements.txt
Django==2.0
pycryptodomex==3.4.7